9809: updated test; moved into reports folders
