+++ web-api/src/reports/getOpinionsFiledByJudgeLambda.ts
