+++ web-api/src/reports/getTrialSessionsForJudgeActivityReportLambda.ts
